config/security: restructured folder and added security config to prepare for login sessionids
