+++ package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "dev": "tsc --watch & nodemon dist/app.js",
+    "dev": "tsc --watch & nodemon dist/src/app.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "keywords": [],
@@ -12,19 +12,29 @@
   "license": "ISC",
   "dependencies": {
     "axios": "^0.19.0",
+    "bluebird": "^3.5.5",
     "body-parser": "^1.19.0",
+    "connect-redis": "^3.4.1",
     "discord.js": "^11.5.1",
     "dotenv": "^8.0.0",
     "express": "^4.17.0",
-    "mongodb": "^3.2.6"
+    "express-session": "^1.16.1",
+    "helmet": "^3.18.0",
+    "mongodb": "^3.2.6",
+    "redis": "^2.8.0"
   },
   "devDependencies": {
+    "@types/bluebird": "^3.5.27",
     "@types/body-parser": "^1.17.0",
+    "@types/connect-redis": "0.0.10",
     "@types/eslint": "^4.16.6",
     "@types/express": "^4.16.1",
+    "@types/express-session": "^1.15.12",
+    "@types/helmet": "0.0.43",
     "@types/jsonwebtoken": "^8.3.2",
     "@types/mongodb": "^3.1.27",
     "@types/node": "^12.0.2",
+    "@types/redis": "^2.8.13",
     "@types/ws": "^6.0.1",
     "@typescript-eslint/parser": "^1.9.0",
     "eslint": "^5.16.0",
