+++ src/routes/oauth/index.ts
@@ -1,0 +1,8 @@
+import express from 'express';
+import routes from './meetup';
+
+const router = express.Router();
+
+router.use('/', routes);
+
+export default router;
